Fix plotting losses diagram after training
Previously, Python would crash when trying to plot the diagram
as it is in a different utils file than the one it was using.

I also made the graph wider and the text bigger to make it easier to read.
It is also saved as a PNG file rather than displayed, which is helpful
when training on a headless system.