Add support for registering images in batches
Previously, each image had to be registered separately. This resulted in some
overhead as the model and reference images/segmentations had to be reloaded
for every registration, despite not having changed.

This commit allows the registration script to load the model and reference files
only once, then re-use them to register every image in a given directory.